Moved MainMenuData to independent header
